fyers-ohlc-backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
pymongo==4.10.1
pandas==2.2.3
numpy==2.1.3
fyers-apiv3==3.4.1
websockets==14.1
requests==2.32.3
